# chore(test): add micro-benchmark performance test for catch operator