--- a/perf/micro/index.js
+++ b/perf/micro/index.js
@@ -14,6 +14,7 @@ Observable.from([
         require("./immediate-scheduler/observable/throw"),
 
         require("./immediate-scheduler/operators/buffer-count"),
+        require("./immediate-scheduler/operators/catch"),
         require("./immediate-scheduler/operators/combine-latest"),
         require("./immediate-scheduler/operators/concat"),
         require("./immediate-scheduler/operators/concat-all"),
@@ -47,6 +48,7 @@ Observable.from([
         require("./current-thread-scheduler/observable/throw"),
 
         require("./current-thread-scheduler/operators/buffer-count"),
+        require("./current-thread-scheduler/operators/catch"),
         require("./current-thread-scheduler/operators/combine-latest"),
         require("./current-thread-scheduler/operators/concat"),
         require("./current-thread-scheduler/operators/concat-all"),