Continue to cover #100.

Able to see noticeable slowness on new operator.

>old catch with immediate scheduler x 87,736 ops/sec ±3.78% (83 runs sampled)
>new catch with immediate scheduler x 61,858 ops/sec ±2.73% (79 runs sampled)
>       -29.5% slower than Rx2

>old catch with current thread scheduler x 73,027 ops/sec ±2.87% (78 runs sampled)
>new catch with current thread scheduler x 58,123 ops/sec ±2.95% (82 runs sampled)
>        -20.41% slower than Rx2

(number varies by system running tests, but shows same trend)