@blesh 2015-09-09
Merged with bb66d4b

Thanks @kwonoj! I'll look into why catch is slower. I have a few ideas.