@kwonoj 2015-09-09
Thanks, I'd like to investigate on my end too but since I'm new to codebase, wasn't trivial for me. Please let me know if I can followup on anything.